Did some things to try to get a cube to display.

+++ src/cube.h
@@ -17,7 +17,7 @@ class Cube : public SceneObject
         void Draw();
     
     private:
-        static GLfloat indexArr[36];
+        static GLuint  indexArr[36];
         static bool    cubeSetUp;
         static GLuint  cubeVBO, cubeVAO, cubeEBO;
